Fix(aggregate-paginate)!: test failure

+++ plugins/mongoose-aggregate-paginate-v2/test/index.test.js
@@ -29,9 +29,9 @@ BookSchema.plugin(mongooseAggregatePaginate);
 const Book = mongoose.model("Book", BookSchema);
 
 beforeAll(async () => {
-  await execute("docker run -d -p 27020:27017 mongo:5.0");
+  if (!process.env.GITHUB_ACTIONS) await execute("docker run -d -p 27017:27017 mongo:5.0");
   await new Promise((resolve) => setTimeout(resolve, 3000));
-  await mongoose.connect("mongodb://localhost:27020/test");
+  await mongoose.connect("mongodb://localhost:27017/test");
   let book,
     books = [];
   const date = new Date();
@@ -52,8 +52,10 @@ beforeAll(async () => {
 
 afterAll(async () => {
   await mongoose.disconnect();
-  await execute("docker stop $(docker ps -q)");
-  await execute("docker rm $(docker ps -aq)");
+  if (!process.env.GITHUB_ACTIONS) {
+    await execute("docker stop $(docker ps -q)");
+    await execute("docker rm $(docker ps -aq)");
+  }
 });
 
 describe("mongoose-paginate", function () {
